Hydrological models

The following rainfall-runoff models are currently included in the package. They can be combined with different snow models.

GR4J

For details about this model, see the following publication:

Perrin, Charles, Claude Michel, and Vazken Andréassian. 2003. “Improvement of a Parsimonious Model for Streamflow Simulation.” Journal of Hydrology 279 (1-4): 275–89. doi:10.1016/S0022-1694(03)00225-7.

Vann.Gr4j — Type.

The Gr4j type contains the state variables (st, st_uh1, st_uh2), the inputs (epot, infilt) for one time step, the parameters (param) and the time step length (tstep) for the GR4J model.

source

The following constructors are available for generating the types:

Vann.Gr4j — Method.

Gr4j(tstep)

Constructor for GR4J with predefined state variables, parameters and inputs. The time step (tstep) is given as a fraction of one day. Thus, for hourly input data tstep should be set to 1/24.

source

Vann.Gr4j — Method.

Gr4j(tstep, param)

Constructor for GR4J with predefined state variables and inputs. The parameter values are given as input. The time step (tstep) is given as a fraction of one day. Thus, for hourly input data tstep should be set to 1/24.

source

The following functions are mainly used during the calibration of the model:

Vann.init_states — Method.

init_states(mdata::Gr4j)

Initilize the state variables of the model.

source

Vann.get_param_range — Method.

get_param_range(mdata::Gr4j)

Get allowed parameter ranges for the calibration of the model.

source

Vann.assign_param — Method.

assign_param(mdata::Gr4j, param::Array{Float64,1})

Assign parameter values to the Gr4j type.

source

The models are written in state-space form. Calling the function below runs the model for one time step.

Vann.hydro_model — Method.

hydro_model(mdata::Gr4j)

Propagate the model one time step and return simulated dischage.

source

HBV