Snow models

The following snow models are currently included in the package. The can be combined with any of the hydrological models.

TinBasic

A basic implementation of a temperature index snow model.

Vann.TinBasic — Type.

The TinBasic type contains the state variables (swe), the inputs (prec, tair) for one time step, the parameters (param) and the time step length (tstep) for a basic temperature index model.

source

The following constructors are available for generating the types:

Vann.TinBasic — Method.

TinBasic(tstep, frac)

Constructor for TinBasic with predefined state variables, parameters and inputs. The time step (tstep) is given as a fraction of one day. Thus, for hourly input data tstep should be set to 1/24. The fraction of elevation bands should sum up to unity.

source

Vann.TinBasic — Method.

TinBasic(tstep, param, frac)

Constructor for TinBasic with predefined state variables and inputs. The time step (tstep) is given as a fraction of one day. Thus, for hourly input data tstep should be set to 1/24. The fraction of elevation bands should sum up to unity.

source

The following functions are mainly used during the calibration of the model:

Vann.init_states — Method.

init_states(mdata::TinBasic)

Initilize the state variables of the model.

source

Vann.get_param_range — Method.

get_param_range(mdata::TinBasic)

Get allowed parameter ranges for the calibration of the model.

source

Vann.assign_param — Method.

assign_param(mdata::TinBasic, param::Array{Float64,1})

Assign parameter values to the TinBasic type.

source

The models are written in state-space form. Calling the function below runs the model for one time step.

snow_model(mdata::TinBasic)

TinStandard

A an enhanced implementation of a temperature index snow model including a liquid water content.

TinStandard

The following constructors are available for generating the types:

TinStandard(tstep, frac)
TinStandard(tstep, param, frac)

The following functions are mainly used during the calibration of the model:

init_states(mdata::TinStandard)
get_param_range(mdata::TinStandard)
assign_param(mdata::TinStandard, param::Array{Float64,1})

The models are written in state-space form. Calling the function below runs the model for one time step.

snow_model(mdata::TinStandard)